_count_type_cards method was move to test class. take_card method and shuffle generated cards added
