+++ test_suite.py
@@ -1,5 +1,6 @@
 import unittest
 from uno.test_uno import TestUno
+from uno.test_stack_uno import TestStack
 from craps.test_craps import TestCraps
 from craps.test_turn import TestTurn
 from craps.test_bets import TestBets
@@ -19,6 +20,7 @@ def suite():
     test_suite = unittest.TestSuite()
     # UNO
     test_suite.addTest(unittest.makeSuite(TestUno))
+    test_suite.addTest(unittest.makeSuite(TestStack))
     # CRAPS
     test_suite.addTest(unittest.makeSuite(TestCraps))
     test_suite.addTest(unittest.makeSuite(TestTurn))
